still working on gettting edit to work
The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link.

Solved:  with using href=!# or something - find stack overflow link
————————————————————————————————————————————————

 //[props.currentId, props.reportObjects]
VScode message:
React Hook useEffect has a missing dependency: 'initialFieldValues'. Either include it or remove the dependency array.

Console warning :
A component is changing a controlled input to be uncontrolled. This is likely caused by the value changing from a defined to undefined, which should not happen. Decide between using a controlled or uncontrolled input element for the lifetime of the component.

————————————————————————————————————————————————

  //[initialFieldValues, props.currentId, props.reportObjects]

VScode message:
The 'initialFieldValues' object makes the dependencies of useEffect Hook (at line 22) change on every render. To fix this, wrap the initialization of 'initialFieldValues' in its own useMemo() Hook.

Console warning :
Maximum update depth exceeded. This can happen when a component calls setState inside useEffect, but useEffect either doesn't have a dependency array, or one of the dependencies changes on every render.

————————————————————————————————————————————————